+++ NEWS.rst
@@ -3,6 +3,10 @@ News
 
 Lists of changes between versions.
 
+Latest
+------
+* Major: Check if linters are installed instead of looking in pyproject.toml.
+
 0.2.0
 -----
 * Improved testing and docstrings